+++ apps/back-office/src/app/application/pages/channels/channels.component.ts
@@ -22,9 +22,13 @@ import { takeUntil } from 'rxjs/operators';
 })
 export class ChannelsComponent extends UnsubscribeComponent implements OnInit {
   // === DATA ===
+  /** Channels list */
   private channels: Channel[] = [];
+  /** Channels data */
   public channelsData: ChannelDisplay[] = [];
+  /** Loading state */
   public loading = true;
+  /** Table columns */
   public displayedColumns: string[] = ['title', 'subscribedRoles', 'actions'];
 
   /**