+++ libs/shared/src/lib/components/widgets/map-settings/graphql/queries.ts
@@ -38,12 +38,25 @@ export const GET_RESOURCE = gql`
 
 /** Get ref data gql query definition */
 export const GET_REFERENCE_DATA = gql`
-  query GetReferenceData($id: ID!) {
+  query GetReferenceData($id: ID!, $aggregation: [ID!]) {
     referenceData(id: $id) {
       id
       name
+      graphQLTypeName
       type
       fields
+      aggregations(ids: $aggregation) {
+        edges {
+          node {
+            id
+            name
+            sourceFields
+            pipeline
+            createdAt
+          }
+        }
+        totalCount
+      }
     }
   }
 `;