+++ libs/ui/src/lib/date/date-picker.directive.spec.ts
@@ -53,6 +53,7 @@ import {
   ],
 })
 class TestingComponent {
+  /** Form */
   form = new FormGroup({
     startDate: new FormControl(''),
     endDate: new FormControl(''),